feat, refactor: Dialog 패키징에 정리, 불필요한 styles 삭제, PostEditConfirmDialog 임시 생성
