+++ app/src/main/res/values/styles.xml
@@ -11,11 +11,6 @@
         <item name="android:stateListAnimator">@null</item>
     </style>
 
-    <style name="CustomSpinner" parent="Widget.AppCompat.Spinner">
-        <item name="android:drawableTint">@color/gray_gray1</item>
-        <item name="android:backgroundTint">@color/gray_gray1</item>
-    </style>
-
     <style name="UserProfileTabTextAppearance" parent="TextAppearance.Design.Tab">
         <item name="android:textSize">16sp</item>
         <item name="android:fontFamily">@font/pretendard_semi_bold</item>
